+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/Renderer.java
@@ -155,6 +155,15 @@ public interface Renderer extends PlayerMessage.Target {
   long DEFAULT_DURATION_TO_PROGRESS_US = 10_000L;
 
   /**
+   * Default duration to progress reported if the renderer does not need calls to {@link #render} to
+   * progress, if it's in {@link #STATE_ENABLED} and already {@linkplain #isReady() ready} or
+   * {@linkplain #isEnded() ended}.
+   *
+   * @see #getDurationToProgressUs
+   */
+  long DEFAULT_IDLE_DURATION_TO_PROGRESS_US = 1_000_000L;
+
+  /**
    * Some renderers can signal when {@link #render(long, long)} should be called.
    *
    * <p>That allows the player to sleep until the next wakeup, instead of calling {@link
@@ -543,7 +552,13 @@ public interface Renderer extends PlayerMessage.Target {
    * Returns minimum amount of playback clock time that must pass in order for the {@link #render}
    * call to make progress.
    *
-   * <p>The default return time is {@link #DEFAULT_DURATION_TO_PROGRESS_US}.
+   * <p>This method may be called when the renderer is in the following states: {@link
+   * #STATE_ENABLED}, {@link #STATE_STARTED}.
+   *
+   * <p>The default return time is {@link #DEFAULT_DURATION_TO_PROGRESS_US} if the renderer is in
+   * {@link #STATE_STARTED}, or in {@link #STATE_ENABLED} and not yet {@linkplain #isReady() ready}
+   * or {@linkplain #isEnded() ended}. Otherwise, it returns {@link
+   * #DEFAULT_IDLE_DURATION_TO_PROGRESS_US}.
    *
    * @param positionUs The current render position in microseconds, measured at the start of the
    *     current iteration of the rendering loop.
@@ -553,7 +568,9 @@ public interface Renderer extends PlayerMessage.Target {
    *     progress.
    */
   default long getDurationToProgressUs(long positionUs, long elapsedRealtimeUs) {
-    return DEFAULT_DURATION_TO_PROGRESS_US;
+    return getState() == STATE_ENABLED && (isReady() || isEnded())
+        ? DEFAULT_IDLE_DURATION_TO_PROGRESS_US
+        : DEFAULT_DURATION_TO_PROGRESS_US;
   }
 
   /**